Add files via upload
Initial

Write Python Code. Generate a ClassList of 100 students. Column 1 represents StudentID from 1 to 100. Column 2 to 6 represents marks in Mathematics, Physics, Chemistry, Biology and English respectively. Create marks using random number generator by considering the following conditions.

     Subject       Min. Marks     Max. Marks
Mathematics         40                  100
    Physics              35                  100
  Chemistry            50                  100
    Biology              40                   90
    English               60                   85
Write a program to calculate the total and average for each student. Total must appear as the 7th column and average as the 8th column. Calculate the average of total marks of the class. Column 9 represents the “Overall grade”. If average mark of a student is greater or equal to class average, Overall Grade is 1. If average mark of the student is less than class average, Overall Grade is 2.